Convert config test to TS
